move all models to own files, move convert all functions to member functions
